refactor: reorgonize repository structure
