+++ examples/todoapp/package.json
@@ -17,7 +17,10 @@
     "redux-logger": "3.0.6"
   },
   "devDependencies": {
-    "parcel-bundler": "^1.12.3"
+    "parcel-bundler": "^1.12.3",
+    "@testing-library/jest-dom": "^4.1.0",
+    "@testing-library/react": "^9.1.3",
+    "jest": "^24.9.0"
   },
   "browserslist": {
     "production": [
