+++ lerna.json
@@ -1,0 +1,11 @@
+{
+  "packages": [
+    "packages/*"
+  ],
+  "hoist": true,
+  "version": "independent",
+  "npmClient": "npm",
+  "npmClientArgs": [
+    "--no-package-lock"
+  ]
+}
