+++ package.json
@@ -1,33 +1,14 @@
 {
-  "name": "@reatom/core",
-  "private": false,
-  "version": "1.0.0-rc9",
-  "description": "State manager with a focus of all needs",
-  "source": "src/index.ts",
-  "main": "build/index.js",
-  "umd:main": "build/index.umd.js",
-  "module": "build/index.mjs",
-  "unpkg": "build/index.umd.js",
-  "types": "build/index.d.ts",
+  "private": true,
+  "version": "0.0.0",
   "scripts": {
-    "prepublishOnly": "npm run rebuild && npm run test",
-    "postpublish": "npm run clean",
-    "build": "microbundle",
-    "clean": "rimraf build/",
-    "rebuild": "npm run clean && npm run build",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:debug": "node --inspect-brk node_modules/.bin/jest --runInBand",
-    "bench": "npm run rebuild && npm run bench:now && npm run clean",
-    "bench:now": "ts-node -P src/benchmark/tsconfig.json src/benchmark",
-    "show-perf": "node ./tools/showPerfResults",
-    "prettify": "prettier --write '**/*.(j|t)s'",
-    "prettify:watch": "onchange '**/*.(j|t)s' -- prettier --write {{changed}}",
-    "ts-check": "tsc --noEmit",
-    "ts-check:watch": "npm run ts-check -- --watch",
-    "update": "npx npm-check-updates -u"
+    "postinstall": "npm run bootstrap",
+    "bootstrap": "lerna bootstrap",
+    "build": "lerna run build",
+    "test": "lerna run test",
+    "prettify": "prettier --write '**/*.{j,t}s'",
+    "prettify:watch": "onchange '**/*.{j,t}s' -- prettier --write {{changed}}"
   },
-  "peerDependencies": {},
   "devDependencies": {
     "@babel/cli": "^7.5.5",
     "@babel/core": "^7.5.5",
@@ -35,23 +16,14 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
     "@babel/polyfill": "^7.4.4",
     "@babel/preset-env": "^7.5.5",
-    "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
-    "@testing-library/jest-dom": "^4.1.0",
-    "@testing-library/react": "^9.1.3",
-    "@types/benchmark": "^1.0.31",
     "@types/jest": "^24.0.18",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.3",
     "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
-    "benchmark": "^2.1.4",
-    "classnames": "^2.2.6",
     "cross-env": "^5.2.0",
     "docsify-cli": "^4.3.0",
-    "effector": "^20.1.2",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
     "eslint": "^6.2.2",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-config-prettier": "^6.1.0",
@@ -59,57 +31,12 @@
     "eslint-plugin-jest": "^22.16.0",
     "eslint-plugin-prettier": "^3.1.0",
     "jest": "^24.9.0",
+    "lerna": "^3.16.4",
     "microbundle": "^0.11.0",
     "onchange": "^6.0.0",
-    "peer-deps-externals-webpack-plugin": "^1.0.4",
     "prettier": "^1.18.2",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "redux": "^4.0.4",
-    "reselect": "^4.0.0",
     "rimraf": "^3.0.0",
     "ts-node": "^8.3.0",
-    "typescript": "^3.6.2",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.7"
-  },
-  "author": "artalar",
-  "license": "MIT",
-  "readme": "README.md",
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/artalar/reatom.git"
-  },
-  "bugs": {
-    "url": "https://github.com/artalar/reatom/issues"
-  },
-  "homepage": "https://artalar.github.io/reatom",
-  "keywords": [
-    "reatom",
-    "@reatom",
-    "state",
-    "store",
-    "flux",
-    "atom",
-    "reactive",
-    "reactivity",
-    "redux",
-    "effector",
-    "mobx",
-    "state-manager",
-    "state manager"
-  ],
-  "files": [
-    "/README.md",
-    "/LICENSE.md",
-    "/package.json",
-    "/src/index.ts",
-    "/build",
-    "/docs"
-  ],
-  "directories": {
-    "lib": "src",
-    "man": "docs",
-    "doc": "docs"
+    "typescript": "^3.6.2"
   }
 }
