+++ packages/core/package.json
@@ -1,10 +1,8 @@
 {
-  "name": "@reatom/react",
-  "version": "0.1.0",
-  "description": "React bindings for @reatom/core",
-  "homepage": "https://github.com/artalar/reatom/tree/master/packages/react",
-  "repository": "https://github.com/artalar/reatom",
-  "license": "MIT",
+  "name": "@reatom/core",
+  "private": false,
+  "version": "1.0.0-rc9",
+  "description": "State manager with a focus of all needs",
   "source": "src/index.ts",
   "main": "build/index.js",
   "umd:main": "build/index.umd.js",
@@ -17,19 +15,36 @@
     "build": "../../node_modules/.bin/microbundle --external=react,@reatom/core",
     "clean": "../../node_modules/.bin/rimraf build/",
     "rebuild": "npm run clean && npm run build",
-    "test": "../../node_modules/.bin/jest src/__tests__/*",
+    "test": "../../node_modules/.bin/jest",
     "test:watch": "../../node_modules/.bin/jest --watch",
     "test:debug": "node --inspect-brk ../../node_modules/.bin/jest --runInBand",
+    "bench": "npm run rebuild && npm run bench:now && npm run clean",
+    "bench:now": "../../node_modules/.bin/ts-node -P src/benchmark/tsconfig.json src/benchmark",
+    "show-perf": "node ./tools/showPerfResults",
     "prettify": "../../node_modules/.bin/prettier --write '**/*.ts'",
     "prettify:watch": "../../node_modules/.bin/onchange '**/*.{j,t}s' -- prettier --write {{changed}}",
     "ts-check": "../../node_modules/.bin/tsc --noEmit",
     "ts-check:watch": "npm run ts-check -- --watch",
     "update": "npx npm-check-updates -u"
   },
-  "peerDependencies": {
-    "react": "^16.8.0",
-    "@reatom/core": "^1.0.0"
+  "devDependencies": {
+    "cross-env": "^5.2.0",
+    "effector": "^20.1.2",
+    "redux": "^4.0.4",
+    "reselect": "^4.0.0",
+    "ts-node": "^8.3.0"
   },
+  "author": "artalar",
+  "license": "MIT",
+  "readme": "README.md",
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/artalar/reatom.git"
+  },
+  "bugs": {
+    "url": "https://github.com/artalar/reatom/issues"
+  },
+  "homepage": "https://artalar.github.io/reatom",
   "keywords": [
     "state",
     "store",
@@ -42,19 +57,12 @@
     "mobx",
     "state-manager",
     "state manager",
-    "reatom",
-    "react",
-    "hooks"
+    "reatom"
   ],
   "files": [
     "/README.md",
     "/package.json",
-    "/index.ts",
+    "/src/index.ts",
     "/build"
-  ],
-  "devDependencies": {
-    "react": "^16.8.0",
-    "@reatom/core": "^1.0.0-rc9",
-    "@types/react": "^16.8.0"
-  }
+  ]
 }
