+++ packages/core/src/__tests__/diamond.ts
