+++ packages/core/src/__tests__/log.json
