+++ packages/core/src/benchmark/tsconfig.json
