+++ packages/core/tools/snapshot.md
