+++ packages/react/src/__tests__/index.ts
@@ -1,0 +1,3 @@
+describe('@reatom/core', () => {
+  test.skip('todo', () => {})
+})
